A cell itself is now selectable and not just the content
When you set the root collection of the cell content to selectable=false,
the selection fills the whole cell. This is useful if you use a wrapper
to store the content of the cells. Then you can see a difference between
selecting the wrapper and only selecting the value stored in the wrapper.